# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: TopJug API
  version: 0.1.0
  description: Initial API contract for the TopJug MVP backend.
  license:
    name: UNLICENSED
    identifier: UNLICENSED
servers:
  - url: /api/v1
paths:
...
  /me/saved-gyms/{gymId}:
    parameters:
      - { $ref: '#/components/parameters/GymIdPath' }
    put:
      operationId: saveGym
      summary: Save a gym idempotently
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Gym saved }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: unsaveGym
      summary: Remove a saved gym idempotently
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Gym removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /records/{recordId}/shares:
    get:
      operationId: listRecordShares
      summary: List revocable share metadata without exposing raw tokens
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RecordIdPath' }
      responses:
        '200':
          description: Share lifecycle metadata
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties: { data: { type: array, items: { $ref: '#/components/schemas/ShareSummary' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createRecordShare
      summary: Create a revocable public share token and optionally attach an S3 image asset
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RecordIdPath' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateShareRequest' } } }
      responses:
        '201':
          description: Share token returned once
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties: { data: { $ref: '#/components/schemas/CreatedShare' } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: topjug_refresh
  parameters:
    GymIdPath:
      name: gymId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RecordIdPath:
      name: recordId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or not owned by the current user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unauthorized:
      description: Missing, invalid, expired, reused, or revoked token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PayloadTooLarge:
      description: JSON request body exceeds 64KB
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateShareRequest:
      type: object
      additionalProperties: false
      properties:
        mediaAssetId: { type: [string, 'null'], format: uuid }
        expiresAt: { type: [string, 'null'], format: date-time }
    CreatedShare:
      type: object
      required: [id, token, apiPath, publicUrl, status, expiresAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        token: { type: string, readOnly: true }
        apiPath: { type: string }
        publicUrl: { type: [string, 'null'], format: uri }
        status: { type: string, enum: [active] }
        expiresAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    ShareSummary:
      type: object
      required: [id, status, mediaAssetId, expiresAt, revokedAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [active, revoked, expired] }
        mediaAssetId: { type: [string, 'null'], format: uuid }
        expiresAt: { type: [string, 'null'], format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string